WIP: Allow prototypes for keywords:
This is just a quick thing I put together when discussing some
syntax for prefix setups with Aristotle. Just an idea and an
implementation we can discuss.

I don't know how important it is to allow prototypes, but if it
doesn't take anything from us, why not.

This breaks the plugin_syntax.t test - I didn't check why.